requirements.txt
fastapi
uvicorn
streamlit
python-multipart
python-dotenv
openai
langchain
pinecone
langchain-community
PyPDF2
python-docx
tqdm
unstructured
